+++ sdk/pubkey/src/lib.rs
@@ -13,7 +13,7 @@ use bytemuck_derive::{Pod, Zeroable};
 #[cfg(feature = "serde")]
 use serde_derive::{Deserialize, Serialize};
 #[cfg(any(feature = "std", target_arch = "wasm32"))]
-use std::vec::Vec;
+use std::{boxed::Box, string::String, vec::Vec};
 #[cfg(feature = "borsh")]
 use {
     borsh::{BorshDeserialize, BorshSchema, BorshSerialize},
@@ -1114,9 +1114,39 @@ macro_rules! pubkey {
     };
 }
 
+/// New random Pubkey for tests and benchmarks.
+#[cfg(all(feature = "std", not(target_os = "solana")))]
+pub fn new_rand() -> Pubkey {
+    Pubkey::from(rand::random::<[u8; PUBKEY_BYTES]>())
+}
+
+#[cfg(all(feature = "std", not(target_os = "solana")))]
+pub fn write_pubkey_file(outfile: &str, pubkey: Pubkey) -> Result<(), Box<dyn std::error::Error>> {
+    use std::io::Write;
+
+    let printable = std::format!("{pubkey}");
+    let serialized = serde_json::to_string(&printable)?;
+
+    if let Some(outdir) = std::path::Path::new(&outfile).parent() {
+        std::fs::create_dir_all(outdir)?;
+    }
+    let mut f = std::fs::File::create(outfile)?;
+    f.write_all(&serialized.into_bytes())?;
+
+    Ok(())
+}
+
+#[cfg(all(feature = "std", not(target_os = "solana")))]
+pub fn read_pubkey_file(infile: &str) -> Result<Pubkey, Box<dyn std::error::Error>> {
+    let f = std::fs::File::open(infile)?;
+    let printable: String = serde_json::from_reader(f)?;
+
+    Ok(Pubkey::from_str(&printable)?)
+}
+
 #[cfg(test)]
 mod tests {
-    use {super::*, strum::IntoEnumIterator};
+    use {super::*, std::fs::remove_file, strum::IntoEnumIterator};
 
     #[test]
     fn test_new_unique() {
@@ -1393,5 +1423,16 @@ mod tests {
             Pubkey::from_str("9h1HyLCW5dZnBVap8C5egQ9Z6pHyjsh5MNy83iPqqRuq").unwrap(),
             PK
         );
+    }
+
+    #[test]
+    fn test_read_write_pubkey() -> Result<(), Box<dyn std::error::Error>> {
+        let filename = "test_pubkey.json";
+        let pubkey = solana_sdk::pubkey::new_rand();
+        write_pubkey_file(filename, pubkey)?;
+        let read = read_pubkey_file(filename)?;
+        assert_eq!(read, pubkey);
+        remove_file(filename)?;
+        Ok(())
     }
 }
